# dm-api/src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "本文档说明了每个接口返回、接收的字段信息。并且制定相应的规范作为前后端交流的渠道。",
    "version": "2.0",
    "title": "易授微服务RESTful APIs说明文档",
    "contact": {
      "name": "刘祥德",
      "url": "/",
      "email": "luvletterU@njust.edu.cn"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/api/v1/data_sets": {
      "get": {
        "tags": [
          "data-set-collection-controller"
        ],
        "summary": "分页查询数据集",
        "operationId": "getPageableUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "size",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "按某属性排序",
            "required": false,
            "type": "string",
            "default": "collectionId"
          },
          {
            "name": "direction",
            "in": "query",
            "description": "排序方式",
            "required": false,
            "type": "string",
            "default": "DESC"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Page«DataSetCollection»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "data-set-collection-controller"
        ],
        "summary": "创建数据集",
        "operationId": "createUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "collectionDto",
            "description": "数据集DTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CollectionDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataSetCollection"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "data-set-collection-controller"
        ],
        "summary": "更新数据集",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "dataSetCollection",
            "description": "dataSetCollection",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DataSetCollection"
            }
          },
          {
            "in": "body",
            "name": "setCollection",
            "description": "数据集编号",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CollectionDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataSetCollection"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "data-set-collection-controller"
        ],
        "summary": "批量删除数据集",
        "operationId": "deleteBatchUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "collectionIds",
            "description": "数据集编号",
            "required": true,
            "schema": {
              "$ref": "#/definitions/List"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DataSetCollection"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/data_sets/{collectionId}": {
      "get": {
        "tags": [
          "data-set-collection-controller"
        ],
        "summary": "根据id获取数据集",
        "operationId": "getByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "description": "数据集编号",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataSetCollection"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "data-set-collection-controller"
        ],
        "summary": "根据id删除数据集",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "description": "数据集编号",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataSetCollection"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/data_sets/{collectionId}/containers": {
      "get": {
        "tags": [
          "data-set-collection-controller"
        ],
        "summary": "获取所有数据集文件信息",
        "operationId": "getContainerUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "description": "数据集编号",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Page«DataSetContainer»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "data-set-collection-controller"
        ],
        "summary": "上传数据集文件",
        "operationId": "addContainerUsingPOST",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "description": "file",
            "required": true,
            "type": "file"
          },
          {
            "name": "collectionId",
            "in": "path",
            "description": "数据集编号",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataSetContainer"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/v1/groups": {
      "get": {
        "tags": [
          "group-controller"
        ],
        "summary": "分页查询分组列表",
        "operationId": "getPageableUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "size",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "按某属性排序",
            "required": false,
            "type": "string",
            "default": "groupId"
          },
          {
            "name": "direction",
            "in": "query",
            "description": "排序方式",
            "required": false,
            "type": "string",
            "default": "DESC"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Page«DataMiningGroup»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "group-controller"
        ],
        "summary": "创建一个分组",
        "operationId": "createUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "group",
            "description": "group",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DataMiningGroupDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataMiningGroup"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "group-controller"
        ],
        "summary": "更新分组信息",
        "operationId": "updateUsingPUT_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "miningGroup",
            "description": "miningGroup",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DataMiningGroup"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataMiningGroup"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "group-controller"
        ],
        "summary": "删除分组",
        "operationId": "deleteUsingDELETE_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "groupIds",
            "description": "groupIds",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DataMiningGroup"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/students": {
      "get": {
        "tags": [
          "student-controller"
        ],
        "summary": "分页查询学生列表",
        "operationId": "studentsPageableUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "className",
            "in": "query",
            "description": "行政班",
            "required": false,
            "type": "string"
          },
          {
            "name": "profession",
            "in": "query",
            "description": "专业",
            "required": false,
            "type": "string"
          },
          {
            "name": "grade",
            "in": "query",
            "description": "年级",
            "required": false,
            "type": "string"
          },
          {
            "name": "studentId",
            "in": "query",
            "description": "学号前缀模糊查询",
            "required": false,
            "type": "string"
          },
          {
            "name": "studentName",
            "in": "query",
            "description": "姓名",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Page«Student»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "student-controller"
        ],
        "summary": "create",
        "operationId": "createUsingPOST_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "student",
            "description": "student",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/StudentDto"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "student-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "studentDto",
            "description": "studentDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StudentDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/StudentDto"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "student-controller"
        ],
        "summary": "根据学号删除学生",
        "operationId": "deleteBatchUsingDELETE_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "studentIds",
            "description": "studentIds",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Student"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AreaType": {
      "type": "object",
      "properties": {
        "abbreviation": {
          "type": "string"
        },
        "areaId": {
          "type": "integer",
          "format": "int32"
        },
        "chineseName": {
          "type": "string"
        },
        "englishName": {
          "type": "string"
        }
      }
    },
    "AssociatedTask": {
      "type": "object",
      "properties": {
        "chineseName": {
          "type": "string"
        },
        "englishName": {
          "type": "string"
        },
        "typeId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "AttributeCharacteristic": {
      "type": "object",
      "properties": {
        "abbreviation": {
          "type": "string"
        },
        "charId": {
          "type": "integer",
          "format": "int32"
        },
        "chineseName": {
          "type": "string"
        },
        "englishName": {
          "type": "string"
        }
      }
    },
    "CollectionDto": {
      "type": "object",
      "properties": {
        "areaId": {
          "type": "integer",
          "format": "int32"
        },
        "associatedTaskIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "attributeCharIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "collectionId": {
          "type": "string"
        },
        "collectionName": {
          "type": "string"
        },
        "dataSetCharIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "dateDonated": {
          "type": "string",
          "format": "date"
        },
        "descriptionIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          }
        },
        "isMissingValues": {
          "type": "string"
        },
        "isUpdated": {
          "type": "boolean"
        },
        "numberOfAttributes": {
          "type": "integer",
          "format": "int64"
        },
        "numberOfInstances": {
          "type": "integer",
          "format": "int64"
        },
        "numberOfWebHits": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "DataMiningGroup": {
      "type": "object",
      "properties": {
        "arrangementId": {
          "type": "string"
        },
        "builtTime": {
          "type": "string",
          "format": "date-time"
        },
        "dataMiningTask": {
          "$ref": "#/definitions/DataMiningTask"
        },
        "groupBuilder": {
          "$ref": "#/definitions/Teacher"
        },
        "groupId": {
          "type": "string"
        },
        "groupLeader": {
          "$ref": "#/definitions/Student"
        },
        "groupName": {
          "type": "string"
        }
      }
    },
    "DataMiningGroupDto": {
      "type": "object",
      "properties": {
        "arrangementId": {
          "type": "string"
        },
        "builderId": {
          "type": "string"
        },
        "builtTime": {
          "type": "string",
          "format": "date-time"
        },
        "groupId": {
          "type": "string"
        },
        "groupName": {
          "type": "string"
        },
        "leaderId": {
          "type": "string"
        },
        "memberIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "taskId": {
          "type": "string"
        }
      }
    },
...
    "DataSetCollection": {
      "type": "object",
      "properties": {
        "area": {
          "$ref": "#/definitions/AreaType"
        },
        "associatedTasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AssociatedTask"
          }
        },
        "attributeCharacteristics": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AttributeCharacteristic"
          }
        },
        "collectionId": {
          "type": "string"
        },
        "collectionName": {
          "type": "string"
        },
        "dataSetCharacteristics": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DataSetCharacteristic"
          }
        },
        "dataSetFolderPath": {
          "type": "string"
        },
        "dateDonated": {
          "type": "string",
          "format": "date"
        },
        "descriptions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DataSetDescription"
          }
        },
        "isMissingValues": {
          "type": "string"
        },
        "numberOfAttributes": {
          "type": "integer",
          "format": "int64"
        },
        "numberOfInstances": {
          "type": "integer",
          "format": "int64"
        },
        "numberOfWebHits": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "DataSetContainer": {
      "type": "object",
      "properties": {
        "containerId": {
          "type": "string"
        },
        "dataSetCollection": {
          "$ref": "#/definitions/DataSetCollection"
        },
        "description": {
          "type": "string"
        },
        "fileName": {
          "type": "string"
        },
        "filePath": {
          "type": "string"
        },
        "fileType": {
          "type": "string"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "Page«DataMiningGroup»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DataMiningGroup"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "$ref": "#/definitions/Sort"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Page«DataSetCollection»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DataSetCollection"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "$ref": "#/definitions/Sort"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Page«DataSetContainer»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DataSetContainer"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "$ref": "#/definitions/Sort"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Page«Student»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Student"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "$ref": "#/definitions/Sort"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Sort": {
      "type": "object"
    },
    "Student": {
      "type": "object",
      "properties": {
        "accountStatus": {
          "type": "string",
          "enum": [
            "LOCKED",
            "AVAILABLE"
          ]
        },
        "avatar": {
          "type": "string",
          "format": "byte"
        },
        "birthday": {
          "type": "string",
          "format": "date"
        },
        "className": {
          "type": "string"
        },
        "favorite": {
          "$ref": "#/definitions/FavoriteStatus"
        },
        "gender": {
          "type": "string"
        },
        "grade": {
          "type": "string"
        },
        "introduction": {
          "type": "string"
        },
        "lastLoginTime": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "profession": {
          "type": "string"
        },
        "publicSalt": {
          "type": "string"
        },
        "regionCode": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/StudentStatus"
        },
        "studentId": {
          "type": "string"
        },
        "studentName": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "StudentDto": {
      "type": "object",
      "properties": {
        "className": {
          "type": "string"
        },
        "favorite": {
          "$ref": "#/definitions/FavoriteStatus"
        },
        "finishedTaskCount": {
          "type": "integer",
          "format": "int64"
        },
        "grade": {
          "type": "string"
        },
        "profession": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/StudentStatus"
        },
        "studentId": {
          "type": "string"
        },
        "studentName": {
          "type": "string"
        }
      }
    },
...
  }
}